Change default service and actuator ports
Since we have decided to go with gRPC as a default API implementation it's better to change default ports accordingly here rather than override defaults in a lot of places.
Existing REST services will not be affected immediately since the chart version is bumped, but if their deployment configs updated to the latest chart then it will be necessary to override ports.